test: Add pytest plugin for spaCy mocking to pytest configuration
