+++ pytest.ini
@@ -1,6 +1,6 @@
 [pytest]
 testpaths = src/icon_extractor/tests_unit src/icon_extractor/tests_integration src/media_analyzer/tests_unit src/media_analyzer/tests_integration src/audio_icon_matcher/tests_unit src/audio_icon_matcher/tests_integration
-addopts = --import-mode=importlib
+addopts = --import-mode=importlib -p tests.pytest_spacy_mock
 asyncio_mode = auto
 markers =
     integration: marks tests as integration tests (may be slow or require external services)
